python/requirements.txt
transformers==4.47.1
pillow==11.0.0 
